1. refactor  src/test/java/org/jfree/chart/PartitioningTest_261P.java          -> src/test/java/test_261P/PartitioningTest_261P.java 2. add test cases on value type Integer, Long, Float, Double, and boundary tests on the max and min value of these types
